Restructure project and add packaging support
Migrated source code and assets into a src/rca_report/ package, added MANIFEST.in and pyproject.toml for PEP 517/518 packaging, and removed top-level parser.py and pytest-results.xml. Updated README with new usage, installation, and folder structure. Added example .env file and included build artifacts in dist/.
