+++ .env.example
@@ -1,0 +1,1 @@
+OPENAI_API_KEY=your-key-here
