+++ src/pytest_rca_report.egg-info/entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+pytest-rca-report = rca_report.parser:main
