+++ src/pytest_rca_report.egg-info/requires.txt
@@ -1,0 +1,4 @@
+matplotlib
+numpy
+python-dotenv
+openai
