+++ src/pytest_rca_report.egg-info/top_level.txt
@@ -1,0 +1,1 @@
+rca_report
