+++ src/rca_report/assets/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>RCA Report</title>
-    <link rel="stylesheet" href="../static/css/style.css">
+    <link rel="stylesheet" href="./static/css/style.css">
     <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
 </head>
 <body>
@@ -82,7 +82,7 @@
 
     <section id="failure-classification" style="display:none"></section>
 </div>
-<script id="data-json" src="../data/results.json" type="application/json"></script>
-<script src="../static/js/app.js"></script>
+<script id="data-json" src="results.json" type="application/json"></script>
+<script src="./static/js/app.js"></script>
 </body>
 </html>
